BM71
Reimplemented BM71 module.

Module worked on 1st revision prototype board, but somehow it lost his MAC address after some configuration. This is known issue about this module and it will be changed in future.

RN4871-V has the same pinout as BM71 and same recommended design, so it can be replaced without any problems.

In future, board with USB and BT will be separate. For now,to lower the cost of prototype, such arrangement will stay for now.

TO DO:
1. Because of difficulties to get TJA1048 in my country and NXP do not have this IC in sample program anymore, this part will be changed to 2 independent CAN transceivers in near future.
2. Maybe replace J2 connector with fancier one, from e.g TE.
4. PIN description on the bottom side of the board.
5. Split board for USB and BT..